requirements.txt
google-genai>=1.64.0
python-dotenv>=1.2.1
httpx>=0.28.1
flask>=3.1.2
flask-compress>=1.13
gunicorn>=24.1.1
flask-sqlalchemy>=3.1.1
psycopg2-binary>=2.9.11
tenacity>=8.0
alembic>=1.13.0
pytest==9.0.2
